useWatchApproachingをuseTransitionHeaderStateにマージ
